Change markup structure
Switched from section to divs, from headings to buttons,
because it makes the implementation of collapsible 
elements easier.

It also helps to avoid html issues during validation
if it is decided to get rid of 'heading' elements inside 
sections, which should be placed inside of them 
according to the specification.